Fix: 2022-03-31

+++ xprocess.cpp
@@ -1424,7 +1424,7 @@ QString XProcess::getLastErrorAsString()
         size_t size=FormatMessageW(FORMAT_MESSAGE_ALLOCATE_BUFFER|FORMAT_MESSAGE_FROM_SYSTEM|FORMAT_MESSAGE_IGNORE_INSERTS,
                                      NULL,nLastError,MAKELANGID(LANG_NEUTRAL,SUBLANG_DEFAULT),(LPWSTR)&messageBuffer,0,NULL);
 
-        sResult=QString::fromWCharArray(messageBuffer,size);
+        sResult=QString::fromWCharArray(messageBuffer,(qint32)size);
 
         LocalFree(messageBuffer);
     }
